src/collectors.py
# ================================================================================================
# COLETORES DE DADOS
# ================================================================================================

import os
import warnings
import requests
import pandas as pd
import numpy as np
from datetime import datetime, timedelta
from typing import Dict, List, Tuple, Optional
import json
import time
import traceback
from bs4 import BeautifulSoup

# Suprimir avisos desnecessários
warnings.filterwarnings('ignore')

class Config:
    """Configurações globais do sistema com APIs reais"""
    
    # Caminhos
    DATA_RAW_PATH = "data/raw/"
    DATA_PROCESSED_PATH = "data/processed/"
    MODELS_PATH = "models/"
    
    # URLs das APIs REAIS
    SINAN_URL = "http://tabnet.datasus.gov.br/cgi/tabcgi.exe"
    INPE_URL = "https://bancodedados.cptec.inpe.br/climatologia/"
    CEMADEN_URL = "http://sjc.salvar.cemaden.gov.br/services/"
    IBGE_URL = "https://servicodados.ibge.gov.br/api/v1/"
    OPENDATA_SUS_URL = "https://opendatasus.saude.gov.br/api/"
    
    # Códigos IBGE
    TEOFILO_OTONI_CODE = "3168606"  # Código IBGE de Teófilo Otoni
    MINAS_GERAIS_CODE = "31"       # Código do estado MG
    
    # Parâmetros do modelo
    SEQUENCE_LENGTH = 52
    PREDICTION_HORIZON = 4
    HIDDEN_DIM = 128
    NUM_HEADS = 8
    NUM_LAYERS = 6
    DROPOUT_RATE = 0.1
    
    # Pesos das doenças na função de perda
    DISEASE_WEIGHTS = {
        'dengue': 0.4,
        'zika': 0.2,
        'chikungunya': 0.2,
        'febre_amarela': 0.2
    }
    
    # Coordenadas de Teófilo Otoni
    TEOFILO_OTONI_LAT = -17.8644
    TEOFILO_OTONI_LON = -41.5056
    
    # Database (PostgreSQL - Render)
    DATABASE_URL = os.getenv("DATABASE_URL", "postgresql://arb_user:14Px7q3HIIqfV4Zq2XCiMjiJuL4N2U0i@dpg-d1nd7nffte5s73f6o410-a/arboviroses_db_pkgd")

class SINANCollector:
    """Coleta dados epidemiológicos reais do SINAN/DataSUS"""
    
    def __init__(self):
        self.base_url = Config.SINAN_URL
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        })
    
    def collect_dengue_data(self, year_start: int = 2015, year_end: int = 2024) -> pd.DataFrame:
        """Coleta dados reais de dengue do SINAN"""
        # Implementação completa aqui
        pass
    
    # Outros métodos de coleta (zika, chikungunya, febre amarela)

class ClimateDataCollector:
    """Coleta dados climáticos reais do INPE/CEMADEN"""
    
    def __init__(self):
        self.inpe_url = Config.INPE_URL
        self.cemaden_url = Config.CEMADEN_URL
        self.session = requests.Session()
    
    def collect_inpe_data(self, start_date: str, end_date: str) -> pd.DataFrame:
        """Coleta dados climáticos do INPE"""
        # Implementação completa aqui
        pass
    
    # Outros métodos de coleta climática

class IBGECollector:
    """Coleta dados demográficos e socioeconômicos do IBGE"""
    
    def __init__(self):
        self.base_url = Config.IBGE_URL
        self.session = requests.Session()
    
    def collect_population_data(self) -> dict:
        """Coleta dados populacionais do município"""
        # Implementação completa aqui
        pass
    
    # Outros métodos de coleta do IBGE

class RealDataCollector:
    """Classe principal para coleta de dados reais"""
    
    def __init__(self):
        self.sinan_collector = SINANCollector()
        self.climate_collector = ClimateDataCollector()
        self.ibge_collector = IBGECollector()
        
    def collect_all_epidemiological_data(self, start_year: int = 2015, end_year: int = 2024) -> pd.DataFrame:
        """Coleta todos os dados epidemiológicos reais"""
        # Implementação completa aqui
        pass
    
    # Outros métodos de coleta integrada
