src/database.py
# ================================================================================================
# GERENCIADOR DE BANCO DE DADOS
# ================================================================================================

import os
import pandas as pd
from sqlalchemy import create_engine, text
from dotenv import load_dotenv

# Carregar variáveis de ambiente
load_dotenv()

class DatabaseManager:
    """Gerenciador de banco de dados PostgreSQL real"""
    
    def __init__(self):
        self.database_url = Config.DATABASE_URL
        self.engine = None
        self.connection = None
    
    def connect(self):
        """Conecta ao banco PostgreSQL"""
        # Implementação completa aqui
        pass
    
    # Outros métodos de gerenciamento de banco de dados
