requirements.txt
pandas
numpy
requests
beautifulsoup4
scikit-learn
sqlalchemy
psycopg2-binary
python-dotenv
matplotlib
seaborn
tqdm
